no way this works but lets try on board

+++ ardunio/serial_test.ino
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <HardwareSerial.h>
 // #include "../shared/CommunicationsConstants.h"]
-#include <cstdint>
+
 
 void setup()
 {
@@ -12,7 +12,7 @@ void loop()
 {
     if (Serial.available() > 0)
     {
-        uint8_t bt = Serial.read();
+        int bt = Serial.read();
 
         Serial.write(bt == 0 ? 8 : bt == 1 ? 34 : 10);
     }
